Flytter gammel logikk ut i egen fil for å minske kaos. IKKE_YRKESAKTIV får tom flyt istedenfor sed flyt - alltid.
